Add test for accepting template defaults
This works already in production, but not in the tests because the
generated Terraform was missing quotes around the default value, which
must have been breaking the parser because it was returning no
parameters.  Then we were removing all non-matching parameters from the
request, which was all of them since the parser returned none.