fix: handle scenario where provisionerdserver deletes task before coderd
Closes https://github.com/coder/internal/issues/1112

We introduced to provisionerdserver the ability to delete a workspace's
associated task for the scenario where a user decided to delete a task
workspace directly, instead of the task. This introduced a race between
coderd and provisionerdserver. We accounted for this in
provisionerdserver with an explicit `!errors.Is(err, sql.ErrNoRows)` but
forgot to apply the same logic to coderd.

In the real world this will almost never happen but it does sometimes
happen during our tests.